Portal 2 behavior gaps: stats controller, level connection, detectors, turret, camera, boss, detail levels, projected texture; Box3D polyhedron fix; portal fast path single-threaded (WIP)

+++ game/shared/portal/portal_usermessages.cpp
@@ -49,7 +49,7 @@ void RegisterUserMessages()
 
 
 	//new stuff for portal
-	usermessages->Register( "EntityPortalled", sizeof( long ) + sizeof( long ) + sizeof( Vector ) + sizeof( QAngle ) ); //something got teleported through a portal
+	usermessages->Register( "EntityPortalled", sizeof( int32 ) + sizeof( int32 ) + sizeof( Vector ) + sizeof( QAngle ) ); //something got teleported through a portal
 	usermessages->Register( "KillCam", -1 );
 
 	// Voting
